refactor: split model_fuser's private helpers into a _private subpackage
Internal's shared/utils/model_fuser encapsulates every underscore-prefixed
helper in _private/fuse.py (restricted visibility), keeping fuse.py's public
surface limited to the 6 fusing-API functions. OSS had everything -- public
API and all 23 private helpers, including the full ONNX export machinery
(dynamo/legacy fallback, IO shape normalization, MHA fastpath handling) --
flattened into one file with no privacy boundary.

Moved all 23 underscore-prefixed functions/classes into a new
model_fuser/_private/ subpackage, mirroring internal's split exactly. Public
fuse.py now only defines fuse_models_to_torchscript/onnx/python,
compute_python_fuse_metadata, get_predictor_output_field_order, and
build_fused_sample_data, importing what they need from _private.fuse.

Updated fuse_test.py's imports and mock.patch targets accordingly (a few
first-pass automated substitutions were wrong -- e.g. a log-message patch
that belonged to the public module, not private -- caught by running the
suite, not assumed correct).

No behavior change: torch/assembler.py, the only external caller, already
only used the 5 public functions that remain in fuse.py.
