+++ python/michelangelo/lib/shared/utils/model_fuser/_private/__init__.py
@@ -1,0 +1,51 @@
+from .fuse import (
+    _align_predictor_input_keys,
+    _build_fused_model_and_sample,
+    _build_fused_sample_input,
+    _build_tx_hydra_spec,
+    _disable_transformer_encoder_fused_fastpath_for_onnx,
+    _expand_batch_for_onnx_export,
+    _export_fused_onnx,
+    _FusedOnnxDynamoTupleWrapper,
+    _FusedOnnxTupleWrapper,
+    _force_onnx_io_shapes_from_schema,
+    _forward_accepts_dict,
+    _forward_param_order,
+    _is_state_dict,
+    _load_module_from_path,
+    _onnx_dynamo_dynamic_shapes_for_tuple_arg,
+    _onnx_dynamo_export_error_should_retry_legacy,
+    _onnx_dynamo_exporter_dependencies_available,
+    _onnx_export_attach_inputs_to_output,
+    _onnx_export_input_preserver,
+    _pl_jit_scripting_guard,
+    _run_export_with_retry,
+    _schema_input_keys,
+    _schema_output_keys,
+)
+
+__all__ = [
+    "_FusedOnnxDynamoTupleWrapper",
+    "_FusedOnnxTupleWrapper",
+    "_align_predictor_input_keys",
+    "_build_fused_model_and_sample",
+    "_build_fused_sample_input",
+    "_build_tx_hydra_spec",
+    "_disable_transformer_encoder_fused_fastpath_for_onnx",
+    "_expand_batch_for_onnx_export",
+    "_export_fused_onnx",
+    "_force_onnx_io_shapes_from_schema",
+    "_forward_accepts_dict",
+    "_forward_param_order",
+    "_is_state_dict",
+    "_load_module_from_path",
+    "_onnx_dynamo_dynamic_shapes_for_tuple_arg",
+    "_onnx_dynamo_export_error_should_retry_legacy",
+    "_onnx_dynamo_exporter_dependencies_available",
+    "_onnx_export_attach_inputs_to_output",
+    "_onnx_export_input_preserver",
+    "_pl_jit_scripting_guard",
+    "_run_export_with_retry",
+    "_schema_input_keys",
+    "_schema_output_keys",
+]
